+++ cmd/hlf/main.go
@@ -18,11 +18,11 @@ import (
 
 var (
 	gFlags = flag.NewFlagsBuilder(
-		flag.NewFlagBuilder("v", "version").
+		flag.NewFlagBuilder("-v", "--version").
 			WithDescription("print information about service"),
-		flag.NewFlagBuilder("h", "help").
+		flag.NewFlagBuilder("-h", "--help").
 			WithDescription("print version of service"),
-		flag.NewFlagBuilder("p", "path").
+		flag.NewFlagBuilder("-p", "--path").
 			WithDescription("set path to config, database files").
 			WithDefaultValue("."),
 	).Build()
@@ -34,12 +34,12 @@ func main() {
 		panic("args invalid")
 	}
 
-	if gFlags.Get("version").GetBoolValue(args) {
+	if gFlags.Get("-v").GetBoolValue(args) {
 		fmt.Println(build.GVersion)
 		return
 	}
 
-	if gFlags.Get("help").GetBoolValue(args) {
+	if gFlags.Get("-h").GetBoolValue(args) {
 		help.Println(settings.GServiceName, settings.CServiceDescription, gFlags)
 		return
 	}