+++ internal/utils/flag/flags_test.go
@@ -6,9 +6,9 @@ import (
 
 var (
 	tgFlags = NewFlagsBuilder(
-		NewFlagBuilder("v", "version").
+		NewFlagBuilder("-v", "--version").
 			WithDescription("print information about service"),
-		NewFlagBuilder("p", "path").
+		NewFlagBuilder("-p", "--path").
 			WithDescription("set path to config, database files").
 			WithDefaultValue("."),
 	).Build()
@@ -25,9 +25,9 @@ func TestPanicFlagsBuilder(t *testing.T) {
 	}()
 
 	_ = NewFlagsBuilder(
-		NewFlagBuilder("v", "version").
+		NewFlagBuilder("-v", "--version").
 			WithDescription("print information about service"),
-		NewFlagBuilder("v", "version").
+		NewFlagBuilder("-v", "--version").
 			WithDescription("print information about service"),
 	).Build()
 }
@@ -35,19 +35,19 @@ func TestPanicFlagsBuilder(t *testing.T) {
 func TestFlagsValidate(t *testing.T) {
 	t.Parallel()
 
-	if ok := tgFlags.Validate([]string{"p"}); ok {
+	if ok := tgFlags.Validate([]string{"-p"}); ok {
 		t.Error("success with void string value")
 		return
 	}
-	if ok := tgFlags.Validate([]string{"q"}); ok {
+	if ok := tgFlags.Validate([]string{"-q"}); ok {
 		t.Error("success with not found flag")
 		return
 	}
-	if ok := tgFlags.Validate([]string{"p", "."}); !ok {
+	if ok := tgFlags.Validate([]string{"-p", "."}); !ok {
 		t.Error("failed with success string flag")
 		return
 	}
-	if ok := tgFlags.Validate([]string{"v"}); !ok {
+	if ok := tgFlags.Validate([]string{"-v"}); !ok {
 		t.Error("failed with success bool flag")
 		return
 	}