+++ internal/utils/help/help_test.go
@@ -10,17 +10,17 @@ func ExamplePrintln() {
 		name.LoadServiceName("hidden-lake-adapter=common"),
 		"adapts HL traffic to a custom HTTP server",
 		flag.NewFlagsBuilder(
-			flag.NewFlagBuilder("v", "version").
+			flag.NewFlagBuilder("-v", "--version").
 				WithDescription("print information about service"),
-			flag.NewFlagBuilder("h", "help").
+			flag.NewFlagBuilder("-h", "--help").
 				WithDescription("print version of service"),
-			flag.NewFlagBuilder("p", "path").
+			flag.NewFlagBuilder("-p", "--path").
 				WithDescription("set path to config, database files").
 				WithDefaultValue("."),
-			flag.NewFlagBuilder("n", "network").
+			flag.NewFlagBuilder("-n", "--network").
 				WithDescription("set network key for connections").
 				WithDefaultValue(""),
-			flag.NewFlagBuilder("t", "threads").
+			flag.NewFlagBuilder("-t", "--threads").
 				WithDescription("set num of parallel functions to calculate PoW").
 				WithDefaultValue("1"),
 		).Build(),
@@ -29,9 +29,9 @@ func ExamplePrintln() {
 	// <Hidden Lake Adapter = Common (HLA=common)>
 	// Description: adapts HL traffic to a custom HTTP server
 	// Arguments:
-	// [ -v, -version ] = print information about service
-	// [ -h, -help ] = print version of service
-	// [ -p, -path ] = set path to config, database files
-	// [ -n, -network ] = set network key for connections
-	// [ -t, -threads ] = set num of parallel functions to calculate PoW
+	// [ -v, --version ] = print information about service
+	// [ -h, --help ] = print version of service
+	// [ -p, --path ] = set path to config, database files
+	// [ -n, --network ] = set network key for connections
+	// [ -t, --threads ] = set num of parallel functions to calculate PoW
 }